fixup! sys/stdio: Introduce helper stdio_write_all
Thanks @maribu for the vibe check

See-Also: https://github.com/RIOT-OS/RIOT/pull/20364#discussion_r1500550941

+++ sys/include/stdio_base.h
@@ -170,7 +170,7 @@ ssize_t stdio_write(const void* buffer, size_t len);
  * This convenience function calls stdio_write as long as not all bytes have
  * sent to @ref stdio_write, or until an error was returned.
  */
-ssize_t stdio_write_all(const void* buffer, size_t len);
+int stdio_write_all(const void* buffer, size_t len);
 
 /**
  * @brief Disable stdio and detach stdio providers
